requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2
django-oauth-toolkit==1.2.0
django-oauth2-provider==0.2.6.1
djangorestframework==3.9.2
idna==2.8
oauthlib==3.0.1
pytz==2018.9
requests==2.21.0
shortuuid==0.5.0
sqlparse==0.3.0
urllib3==1.24.1
